+++ scripts/lazy.py
@@ -16,8 +16,15 @@ def getMainLLMFilenamePath(path: str, extension="v"):
 
     return os.path.join(path, f"{filename}.{extension}")
 
-def readFileContent(filePath:str, mode='r+'):
+
+def readFileContent(filePath: str, mode="r+"):
     file = open(filePath, mode)
-    fileContent:str = file.read()
+    fileContent: str = file.read()
     file.close()
     return fileContent
+
+
+def saveFileContent(filePath: str, content: str, mode="w+"):
+    file = open(filePath, mode)
+    file.write(content)
+    file.close()